langchain-client.py
import asyncio

from dotenv import load_dotenv
from langchain_mcp_adapters.client import MultiServerMCPClient
from langchain_openai import ChatOpenAI
from langgraph.prebuilt import create_react_agent

load_dotenv()

llm = ChatOpenAI(model="gpt-4o-mini")

async def main():
    print("Hello from mcp-practise!")

if __name__ == "__main__":
    asyncio.run(main())

