pyproject.toml
[project]
name = "mcp-practise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.9",
    "python-dotenv>=1.2.2",
]
